third_party/WebKit/LayoutTests/editing/inserting/insert-table-in-paragraph-crash.html

<div id="container">
Editable:<div contenteditable="true" id="editable"></div>
</div>
</div>
<script src="../../resources/js-test.js"></script>
<script>
function $(id) { return document.getElementById(id); }
$('editable').focus();
var expected = '<table><tbody><tr><td>stats</td></tr></tbody></table>';
document.execCommand('InsertHTML', false, '<p>' + expected + '</p>');
shouldBeEqualToString('$("editable").innerHTML', expected);
if (window.testRunner)
    $('container').outerHTML = '';
</script>
</body>
